requirements.txt
beautifulsoup4
python-telegram-bot
requests
tweepy
